Seperate interface into seperate file
