+++ TGServiceInterface/Server.cs
@@ -1,8 +1,6 @@
-﻿using RGiesecke.DllExport;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Net;
-using System.Runtime.InteropServices;
 using System.ServiceModel;
 namespace TGServiceInterface
 {
@@ -146,43 +144,6 @@ namespace TGServiceInterface
 			{
 				return false;
 			}
-		}
-	}
-
-	/// <summary>
-	/// Used by DD to access the interop API with call()()
-	/// </summary>
-	[ServiceContract]
-	public interface ITGServiceBridge
-	{
-		/// <summary>
-		/// Called from /world/ExportService(command)
-		/// </summary>
-		/// <param name="command">The command to run</param>
-		/// <returns>true on success, false on failure</returns>
-		[OperationContract]
-		bool InteropMessage(string command);
-	}
-
-	/// <summary>
-	/// Holds the proc that DD calls to access <see cref="ITGServiceBridge"/>
-	/// </summary>
-	public class DDInteropCallHolder
-	{
-		/// <summary>
-		/// The proc that DD calls to access <see cref="ITGServiceBridge"/>
-		/// </summary>
-		/// <param name="args">The arguments passed</param>
-		/// <returns>0 success, -1 on a WCF, error, 1 on an operation error</returns>
-		[DllExport("DDEntryPoint", CallingConvention = CallingConvention.Cdecl)]
-		public static int DDEntryPoint(int argc, [MarshalAs(UnmanagedType.LPArray, ArraySubType = UnmanagedType.LPStr, SizeParamIndex = 0)]string[] args)
-		{
-			try
-			{
-				Server.GetComponent<ITGServiceBridge>().InteropMessage(String.Join(" ", args));
-			}
-			catch { }
-			return 0;
 		}
 	}
 }
